+++ apps/minifront/src/components/swap/swap-form/token-swap-input.tsx
@@ -41,8 +41,9 @@ const getAssetOutBalance = (
     const balanceViewMetadata = getMetadataFromBalancesResponse(balance);
 
     return (
-      getAddressIndex(balance.accountAddress).equals(getAddressIndex(assetIn.accountAddress)) &&
-      assetOut.equals(balanceViewMetadata)
+      getAddressIndex(balance.accountAddress).account ===
+        getAddressIndex(assetIn.accountAddress).account &&
+      assetOut.penumbraAssetId?.equals(balanceViewMetadata.penumbraAssetId)
     );
   });
   const matchedBalance = getBalanceView.optional()(match);
@@ -161,6 +162,9 @@ export const TokenSwapInput = () => {
             </div>
           </div>
         </div>
+      </div>
+
+      <div className='pt-2'>
         {priceHistory.startMetadata && priceHistory.endMetadata && priceHistory.candles.length ? (
           <CandlestickPlot
             className='h-[480px] w-full bg-charcoal'