+++ packages/storage/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @penumbra-zone/storage
 
+## 13.0.0
+
+### Minor Changes
+
+- fa798d9: Bufbuild + registry dep update
+
+### Patch Changes
+
+- Updated dependencies [fa798d9]
+- Updated dependencies [fa798d9]
+  - @penumbra-zone/wasm@16.0.0
+  - @penumbra-zone/getters@11.0.0
+  - @penumbra-zone/types@15.0.0
+
+## 12.0.0
+
+### Major Changes
+
+- 28a48d7: Fixes for multi-asset fees
+
+### Patch Changes
+
+- Updated dependencies [28a48d7]
+- Updated dependencies [28a48d7]
+  - @penumbra-zone/wasm@15.0.0
+  - @penumbra-zone/types@14.0.0
+
 ## 11.0.0
 
 ### Major Changes