+++ packages/types/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @penumbra-zone/types
 
+## 15.0.0
+
+### Minor Changes
+
+- fa798d9: Bufbuild + registry dep update
+
+### Patch Changes
+
+- Updated dependencies [fa798d9]
+  - @penumbra-zone/getters@11.0.0
+
+## 14.0.0
+
+### Major Changes
+
+- 28a48d7: Fixes for multi-asset fees
+
 ## 13.1.0
 
 ### Minor Changes